+++ CHANGES.md
@@ -1,5 +1,8 @@
 unreleased
-============
+==========
+
+  + merlin library
+    - Fix signature-help with type aliases (#2067, fixes #1927)
   + test suite
     - Remove the FIXME line for #1404 as the issue was already fixed and add two tests (#2073).